reorg non-task methods out of tasks.py to utils.py
